add donors and award calculated props to shared_calc_prop file
