+++ src/encoded/types/suspension.py
@@ -6,6 +6,9 @@ from snovault import (
 from .base import (
     Item,
 )
+from .shared_calculated_properties import (
+    CalculatedDonors,
+)
 
 
 @collection(
@@ -15,7 +18,7 @@ from .base import (
         'title': 'Suspensions',
         'description': 'Listing of Suspensions',
     })
-class Suspension(Item):
+class Suspension(Item, CalculatedDonors):
     item_type = 'suspension'
     schema = load_schema('encoded:schemas/suspension.json')
     embedded = [
@@ -24,24 +27,3 @@ class Suspension(Item):
         'donors.diseases',
         'donors.ethnicity'
     ]
-
-
-    @calculated_property(define=True,
-                         schema={"title": "Donors",
-                                 "description": "The donors the sample was derived from.",
-                                 "comment": "Do not submit. This is a calculated property",
-                                 "type": "array",
-                                 "items": {
-                                    "type": "string",
-                                    "linkTo": "Donor"
-                                    }
-                                })
-    def donors(self, request, derived_from):
-        all_donors = set()
-        for bs in derived_from:
-            bs_obj = request.embed(bs, '@@object')
-            if 'Donor' in bs_obj.get('@type'):
-                all_donors.add(bs_obj.get('@id'))
-            else:
-                all_donors.update(bs_obj.get('donors'))
-        return sorted(all_donors)
